refactor(gamelog): improve callback handling
